Fix logf lines in codecs (type mismatches)
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@26038 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/aac.c
@@ -247,7 +247,7 @@ next_track:
     err = CODEC_OK;
 
 done:
-    LOGF("AAC: Decoded %lu samples\n", sound_samples_done);
+    LOGF("AAC: Decoded %lu samples\n", (unsigned long)sound_samples_done);
 
     if (ci->request_next_track())
         goto next_track;
